+++ test/e2e/utils/nodes/nodes.go
@@ -1,14 +1,30 @@
-package utils
+/*
+Copyright 2020 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package nodes
 
 import (
 	"context"
 	"fmt"
-	"os"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
+
 	"k8s.io/kubernetes/test/e2e/framework"
 )
 
@@ -71,11 +87,4 @@ func FilterNodesWithEnoughCores(nodes []v1.Node, cpuAmount string) ([]v1.Node, e
 	}
 
 	return resNodes, nil
-}
-
-func GetNodeName() string {
-	if nodeName, ok := os.LookupEnv("E2E_WORKER_NODE_NAME"); ok {
-		return nodeName
-	}
-	return DefaultNodeName
 }
