+++ test/e2e/utils/testenv/testenv.go
@@ -1,0 +1,43 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package testenv
+
+import (
+	"os"
+)
+
+const (
+	// we rely on kind for our CI
+	DefaultNodeName  = "kind-worker"
+	DefaultNamespace = "default"
+	RTELabelName     = "resource-topology"
+	RTEContainerName = "resource-topology-exporter-container"
+)
+
+func GetNodeName() string {
+	if nodeName, ok := os.LookupEnv("E2E_WORKER_NODE_NAME"); ok {
+		return nodeName
+	}
+	return DefaultNodeName
+}
+
+func GetNamespaceName() string {
+	if nsName, ok := os.LookupEnv("E2E_NAMESPACE_NAME"); ok {
+		return nsName
+	}
+	return DefaultNamespace
+}
